Refactor a ton of stuff into specific class files
