+++ .gitignore
@@ -1,2 +1,4 @@
 /.idea
 .DS_Store
+src.elf
+build/
