+++ src/source/icons.h
@@ -1,1155 +1,161 @@
-u8 emptyICO[64]={
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+#ifndef ICONS_H
+#define ICONS_H
+
+#include <nds.h>
+#include <maxmod9.h>
+
+extern u8 emptyICO[64];
 //29 darkest, 31 lightest
 //13 dark gold, 7 light gold
-u8 goldICO[64]={
-	1,1,1,1,1,1,1,1,
-	1,1,1,13,13,1,1,1,
-	1,1,13,13,7,13,1,1,
-    1,13,7,13,7,13,13,1,
-	1,13,7,13,7,13,13,1,
-    13,7,13,7,13,7,13,13,
-	13,7,13,7,13,7,13,13,
-    1,1,1,1,1,1,1,1
-};
+extern u8 goldICO[64];
 
-u8 heartFullICO[64]={
-	01,01,01,01,01,01,01,01,
-	01,01,8,01,01,8,01,01,
-	01,8,20,8,8,20,8,01,
-	01,8,20,20,20,20,8,01,
-	01,8,20,20,20,20,8,01,
-	01,01,8,20,20,8,01,01,
-	01,01,01,8,8,01,01,01,
-	01,01,01,01,01,01,01,01
-};
+extern u8 heartFullICO[64];
 
-u8 heartEmptyICO[64]={
-	01,01,01,01,01,01,01,01,
-	01,01,8,01,01,8,01,01,
-	01,8,01,8,8,01,8,01,
-	01,8,01,01,01,01,8,01,
-	01,8,01,01,01,01,8,01,
-	01,01,8,01,01,8,01,01,
-	01,01,01,8,8,01,01,01,
-	01,01,01,01,01,01,01,01
-};
+extern u8 heartEmptyICO[64];
 
-u8 rockICO[64]={
-	01,01,36,36,36,36,01,01,
-	01,36,41,41,41,41,36,01,
-	36,41,41,41,41,40,40,36,
-	36,41,41,40,40,40,36,36,
-	36,40,40,40,36,40,36,36,
-	36,40,40,37,37,36,37,36,
-	36,40,36,37,36,37,37,36,
-	01,36,36,36,36,36,36,01
-};
+extern u8 rockICO[64];
 
-u8 redButtonUpICO[64]={
-	1,1,1,1,1,12,12,1,
-	1,1,1,1,1,12,12,1,
-	25,25,25,25,25,12,12,25,
-	25,23,23,23,23,12,12,25,
-	25,23,23,23,23,8,8,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 redButtonMiddleICO[64]={
-	1,1,1,12,12,1,1,1,
-	1,1,1,12,12,1,1,1,
-	25,25,25,12,12,25,25,25,
-	25,23,23,12,12,23,23,25,
-	25,23,23,8,8,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 redButtonDownICO[64]={
-	1,12,12,1,1,1,1,1,
-	1,12,12,1,1,1,1,1,
-	25,12,12,25,25,25,25,25,
-	25,12,12,23,23,23,23,25,
-	25,8,8,23,23,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 redButtonUpICO[64];
+extern u8 redButtonMiddleICO[64];
+extern u8 redButtonDownICO[64];
 
-u8 doorDownICO[64]={
-    1,1,1,1,1,1,1,1,
-	12,12,12,2,2,12,12,12,
-	12,12,12,2,2,12,12,12,
-	12,12,12,2,2,12,12,12,
-	12,12,12,2,2,12,12,12,
-	12,12,12,2,2,12,12,12,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownBICO[64]={
-    1,1,1,1,1,1,1,1,
-	12,12,12,1,1,12,12,12,
-	12,12,12,1,1,12,12,12,
-	12,12,12,1,1,12,12,12,
-	12,12,12,1,1,12,12,12,
-	12,12,12,1,1,12,12,12,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownCICO[64]={
-    1,1,1,1,1,1,1,1,
-	12,12,1,1,1,1,12,12,
-	12,12,1,1,1,1,12,12,
-	12,12,1,1,1,1,12,12,
-	12,12,1,1,1,1,12,12,
-	12,12,1,1,1,1,12,12,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownDICO[64]={
-    1,1,1,1,1,1,1,1,
-	12,1,1,1,1,1,1,12,
-	12,1,1,1,1,1,1,12,
-    12,1,1,1,1,1,1,12,
-    12,1,1,1,1,1,1,12,
-	12,1,1,1,1,1,1,12,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 doorDownICO[64];
+extern u8 doorDownBICO[64];
+extern u8 doorDownCICO[64];
+extern u8 doorDownDICO[64];
 
-u8 doorLeftICO[64]={
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,2,2,2,1,1,1,1,
-	1,2,2,2,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1
-};
-u8 doorLeftBICO[64]={
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1
-};
-u8 doorLeftCICO[64]={
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,12,12,12,1,1,1,1,
-	1,12,12,12,1,1,1,1
-};
-u8 doorLeftDICO[64]={
-	1,12,12,12,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,12,12,12,1,1,1,1
-};
+extern u8 doorLeftICO[64];
+extern u8 doorLeftBICO[64];
+extern u8 doorLeftCICO[64];
+extern u8 doorLeftDICO[64];
 
-u8 doorRightICO[64]={
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,2,2,2,1,
-	1,1,1,1,2,2,2,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-};
-u8 doorRightBICO[64]={
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-};
-u8 doorRightCICO[64]={
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,12,12,12,1,
-};
-u8 doorRightDICO[64]={
-	1,1,1,1,12,12,12,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,12,12,12,1,
-};
+extern u8 doorRightICO[64];
+extern u8 doorRightBICO[64];
+extern u8 doorRightCICO[64];
+extern u8 doorRightDICO[64];
 
-u8 greenButtonUpICO[64]={
-	1,1,1,1,1,28,28,1,
-	1,1,1,1,1,28,28,1,
-	25,25,25,25,25,28,28,25,
-	25,23,23,23,23,28,28,25,
-	25,23,23,23,23,8,8,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 greenButtonMiddleICO[64]={
-	1,1,1,28,28,1,1,1,
-	1,1,1,28,28,1,1,1,
-	25,25,25,28,28,25,25,25,
-	25,23,23,28,28,23,23,25,
-	25,23,23,8,8,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 greenButtonDownICO[64]={
-	1,28,28,1,1,1,1,1,
-	1,28,28,1,1,1,1,1,
-	25,28,28,25,25,25,25,25,
-	25,28,28,23,23,23,23,25,
-	25,8,8,23,23,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 greenButtonUpICO[64];
+extern u8 greenButtonMiddleICO[64];
+extern u8 greenButtonDownICO[64];
 
 
-u8 doorGDownICO[64]={
-    1,1,1,1,1,1,1,1,
-	28,28,28,32,32,28,28,28,
-	28,28,28,32,32,28,28,28,
-	28,28,28,32,32,28,28,28,
-	28,28,28,32,32,28,28,28,
-	28,28,28,32,32,28,28,28,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorGDownBICO[64]={
-    1,1,1,1,1,1,1,1,
-	28,28,28,1,1,28,28,28,
-	28,28,28,1,1,28,28,28,
-	28,28,28,1,1,28,28,28,
-	28,28,28,1,1,28,28,28,
-	28,28,28,1,1,28,28,28,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorGDownCICO[64]={
-    1,1,1,1,1,1,1,1,
-	28,28,1,1,1,1,28,28,
-	28,28,1,1,1,1,28,28,
-	28,28,1,1,1,1,28,28,
-	28,28,1,1,1,1,28,28,
-	28,28,1,1,1,1,28,28,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorGDownDICO[64]={
-    1,1,1,1,1,1,1,1,
-	28,1,1,1,1,1,1,28,
-	28,1,1,1,1,1,1,28,
-	28,1,1,1,1,1,1,28,
-	28,1,1,1,1,1,1,28,
-	28,1,1,1,1,1,1,28,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 doorGDownICO[64];
+extern u8 doorGDownBICO[64];
+extern u8 doorGDownCICO[64];
+extern u8 doorGDownDICO[64];
 
-u8 doorGLeftICO[64]={
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,32,32,32,1,1,1,1,
-	1,32,32,32,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-};
-u8 doorGLeftBICO[64]={
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-};
-u8 doorGLeftCICO[64]={
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-};
-u8 doorGLeftDICO[64]={
-	1,28,28,28,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,28,28,28,1,1,1,1,
-};
+extern u8 doorGLeftICO[64];
+extern u8 doorGLeftBICO[64];
+extern u8 doorGLeftCICO[64];
+extern u8 doorGLeftDICO[64];
 
-u8 doorGRightICO[64]={
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,32,32,32,1,
-	1,1,1,1,32,32,32,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-};
-u8 doorGRightBICO[64]={
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-};
-u8 doorGRightCICO[64]={
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,28,28,28,1,
-};
-u8 doorGRightDICO[64]={
-	1,1,1,1,28,28,28,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,28,28,28,1,
-};
+extern u8 doorGRightICO[64];
+extern u8 doorGRightBICO[64];
+extern u8 doorGRightCICO[64];
+extern u8 doorGRightDICO[64];
 
-u8 wallICO[64]={
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	19,39,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39,
-	39,39,39,39,19,19,39,39,
-	19,39,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39
-};
-u8 wallDownICO[64]={
-	1,18,18,18,18,18,18,1,
-	18,42,42,02,02,02,02,18,
-	18,42,02,02,42,42,02,18,
-	18,42,42,02,42,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallUpICO[64]={
-	18,42,2,2,2,42,42,18,
-	18,18,2,2,2,2,18,18,
-    18,39,18,18,18,18,39,18,
-	18,19,19,39,19,39,19,18,
-	18,19,19,39,39,39,39,18,
-	18,39,39,39,39,19,19,18,
-	18,19,19,39,19,19,39,18,
-	1,39,39,39,39,39,39,1
-};
-u8 wallRightICO[64]={
-	1,18,18,18,18,18,18,18,
-	18,2,42,2,2,2,42,2,
-	39,18,18,18,18,18,18,18,
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	39,19,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39
+extern u8 wallICO[64];
+extern u8 wallDownICO[64];
+extern u8 wallUpICO[64];
+extern u8 wallRightICO[64];
+extern u8 wallLeftICO[64];
+extern u8 wallUpRightICO[64];
+extern u8 wallUpLeftICO[64];
 
-};
-u8 wallLeftICO[64]={
-	18,18,18,18,18,18,18,1,
-	2,2,42,2,42,2,2,18,
-	18,18,18,18,18,18,18,39,
-	39,19,19,39,19,39,19,39,
-	39,19,19,39,39,39,39,39,
-	39,39,39,39,19,19,39,39,
-	19,39,19,39,19,19,19,39,
-	39,39,39,39,39,39,39,39
-};
-u8 wallUpRightICO[64]={
-	18,2,2,2,2,2,42,2,
-	18,2,2,42,2,2,2,2,
-	39,18,18,18,18,18,18,18,
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	39,19,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39
-};
-u8 wallUpLeftICO[64]={
-	42,2,2,42,2,2,2,18,
-	2,2,2,42,2,2,2,18,
-	18,18,18,18,18,18,18,39,
-	39,19,19,39,19,39,19,39,
-	39,19,19,39,39,39,39,39,
-	39,39,39,39,39,19,19,39,
-	19,39,19,39,19,19,19,39,
-	39,39,39,39,39,39,39,39
-};
+extern u8 wallUpDownICO[64];
 
-u8 wallUpDownICO[64]={
-	18,42,02,42,42,02,42,18,
-	18,42,02,42,02,02,02,18,
-	18,02,02,02,02,42,42,18,
-	18,42,42,42,02,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
+extern u8 wallDownRightICO[64];
+extern u8 wallDownLeftICO[64];
+extern u8 wallLeftRightICO[64];
+extern u8 wallUpLeftDownICO[64];
+extern u8 wallUpRightDownICO[64];
+extern u8 wallUpLeftRightICO[64];
+extern u8 wallDownLeftRightICO[64];
 
-u8 wallDownRightICO[64]={
-	1,18,18,18,18,18,18,18,
-	18,42,42,02,02,02,02,42,
-	18,42,02,02,42,42,02,02,
-	18,42,42,02,42,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallDownLeftICO[64]={
-	18,18,18,18,18,18,18,1,
-	42,42,42,02,02,02,02,18,
-	02,42,02,02,42,42,02,18,
-	18,42,42,02,42,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallLeftRightICO[64]={
-	18,18,18,18,18,18,18,18,
-	2,42,2,2,42,42,2,2,
-	18,18,18,18,18,18,18,18,
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	19,39,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39
-};
-u8 wallUpLeftDownICO[64]={
-	02,42,02,42,42,02,42,18,
-	02,42,02,42,02,02,02,18,
-	02,02,02,02,02,42,42,18,
-	18,42,42,42,02,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallUpRightDownICO[64]={
-	18,42,02,42,42,02,42,02,
-	18,42,02,42,02,02,02,02,
-	18,02,02,02,02,42,42,02,
-	18,42,42,42,02,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallUpLeftRightICO[64]={
-	2,2,42,2,2,2,42,2,
-	2,2,42,2,2,42,42,2,
-	18,18,18,18,18,18,18,18,
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	19,39,19,39,19,19,39,19,
-	39,39,39,39,39,39,39,39
-};
-u8 wallDownLeftRightICO[64]={
-	18,18,18,18,18,18,18,18,
-	42,42,02,42,02,02,02,02,
-	02,02,02,02,02,42,42,02,
-	18,42,42,42,02,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-
-u8 wallUpDownLeftRightICO[64]={
-	02,42,02,42,42,02,42,42,
-	02,42,02,42,02,02,02,02,
-	02,02,02,02,02,42,42,02,
-	18,42,42,42,02,42,42,18,
-	18,02,02,02,02,02,02,18,
-	18,02,42,42,42,02,42,18,
-	18,02,02,42,42,02,02,18,
-	18,42,02,02,02,02,42,18
-};
-u8 wallIslandICO[64]={
-	1,18,18,18,18,18,18,1,
-	18,42,2,2,42,2,2,18,
-	19,18,18,18,18,18,18,19,
-	39,19,19,39,19,39,19,19,
-	39,19,19,39,39,39,39,19,
-	39,39,39,39,19,19,39,39,
-	39,19,19,39,19,19,19,39,
-	39,39,39,39,39,39,39,39
-};
+extern u8 wallUpDownLeftRightICO[64];
+extern u8 wallIslandICO[64];
 
 
-u8 manDownICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,7,7,7,1,23,
-	1,1,7,8,7,8,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,28,27,1,23,
-	1,7,27,27,28,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
-u8 manDownAICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,7,7,7,1,23,
-	1,1,7,8,7,8,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,28,27,1,23,
-	1,7,27,27,28,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,1,1,1
-};
-u8 manDownBICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,7,7,7,1,23,
-	1,1,7,8,7,8,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,28,27,1,23,
-	1,7,27,27,28,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,1,1,1,11,1,1
-};
-u8 manDownAttackICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,7,7,7,1,1,
-	1,1,7,8,7,8,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,28,27,1,23,
-	1,7,27,27,28,27,1,23,
-	1,1,28,28,28,28,28,7,
-	1,1,11,1,1,11,1,1
-};
+extern u8 manDownICO[64];
+extern u8 manDownAICO[64];
+extern u8 manDownBICO[64];
+extern u8 manDownAttackICO[64];
 
-u8 manUpAICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,23,
-	1,1,26,26,26,26,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,27,27,1,23,
-	1,7,27,27,27,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,1,1,1
-};
-u8 manUpBICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,23,
-	1,1,26,26,26,26,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,27,27,1,23,
-	1,7,27,27,27,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,1,1,1,11,1,1
-};
-u8 manUpICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,23,
-	1,1,26,26,26,26,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,27,27,1,23,
-	1,7,27,27,27,27,28,7,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
-u8 manUpAttackICO[64]={
-	1,1,26,26,26,26,1,23,
-	1,1,26,26,26,26,1,23,
-	1,1,26,26,26,26,1,23,
-	1,1,7,7,7,7,1,23,
-	1,1,27,27,27,27,28,7,
-	1,7,27,27,27,27,1,1,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
+extern u8 manUpAICO[64];
+extern u8 manUpBICO[64];
+extern u8 manUpICO[64];
+extern u8 manUpAttackICO[64];
 
 
-u8 manLeftICO[64]={
-	1,1,26,26,26,26,1,1,
-	23,1,7,7,7,26,1,1,
-	23,1,8,7,8,7,1,1,
-	23,1,7,7,7,7,1,1,
-	23,1,27,28,27,27,1,1,
-	7,28,27,28,27,27,7,1,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
-u8 manLeftAICO[64]={
-	1,1,26,26,26,26,1,1,
-	23,1,7,7,7,26,1,1,
-	23,1,8,7,8,7,1,1,
-	23,1,7,7,7,7,1,1,
-	23,1,27,28,27,27,1,1,
-	7,28,27,28,27,27,7,1,
-	1,1,28,28,28,28,1,1,
-	1,1,1,1,1,11,1,1
-};
-u8 manLeftBICO[64]={
-	1,1,26,26,26,26,1,1,
-	23,1,7,7,7,26,1,1,
-	23,1,8,7,8,7,1,1,
-	23,1,7,7,7,7,1,1,
-	23,1,27,28,27,27,1,1,
-	7,28,27,28,27,27,7,1,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,1,1,1
-};
-u8 manLeftAttackICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,23,7,7,7,26,1,1,
-	1,23,8,7,8,7,1,1,
-	1,23,7,7,7,7,1,1,
-	1,23,27,28,27,27,1,1,
-	1,7,27,28,27,27,7,1,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
+extern u8 manLeftICO[64];
+extern u8 manLeftAICO[64];
+extern u8 manLeftBICO[64];
+extern u8 manLeftAttackICO[64];
 
-u8 manRightAttackICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,7,7,7,23,1,
-	1,1,7,8,7,8,23,1,
-	1,1,7,7,7,7,23,1,
-	1,1,27,27,28,27,23,1,
-	1,7,27,27,28,27,7,1,
-	1,1,28,28,28,28,1,1,
-	1,1,11,1,1,11,1,1
-};
+extern u8 manRightAttackICO[64];
 
 
-u8 endDownICO[64]={
-	01,01,25,01,01,25,01,01,
-	01,01,25,25,25,25,01,01,
-	01,01,24,01,01,24,01,01,
-	01,01,24,24,24,24,01,01,
-	01,01,23,8,8,23,01,01,
-	01,8,23,23,23,23,8,01,
-	01,8,22,8,8,22,8,01,
-	01,01,22,22,22,22,01,01
-};
-u8 endDownBICO[64]={
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,1,
-	1,1,7,7,7,7,1,1,
-	1,1,27,27,27,27,1,1,
-	1,7,27,27,27,27,7,1,
-	1,8,28,28,28,28,8,1,
-	1,1,11,8,8,11,1,1
-};
-u8 endDownCICO[64]={
-	1,1,25,1,1,25,1,1,
-	1,1,25,25,25,25,1,1,
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,1,
-	1,1,26,26,26,26,1,1,
-	1,8,7,7,7,7,8,1,
-	1,8,27,27,27,27,8,1,
-	1,1,27,27,27,27,1,1
-};
-u8 endDownDICO[64]={
-	1,1,25,1,1,25,1,1,
-	1,1,25,25,25,25,1,1,
-	1,1,24,1,1,24,1,1,
-	1,1,24,24,24,24,1,1,
-	1,1,26,26,26,26,1,1,
-	1,8,26,26,26,26,8,1,
-	1,8,26,26,26,26,8,1,
-	1,1,7,7,7,7,1,1
-};
-u8 endDownEICO[64]={
-	1,1,25,1,1,25,1,1,
-	1,1,25,25,25,25,1,1,
-	1,1,24,1,1,24,1,1,
-	1,1,24,24,24,24,1,1,
-	1,1,23,8,8,23,1,1,
-	1,8,23,23,23,23,8,1,
-	1,8,26,26,26,26,8,1,
-	1,1,26,26,26,26,1,1
-};
+extern u8 endDownICO[64];
+extern u8 endDownBICO[64];
+extern u8 endDownCICO[64];
+extern u8 endDownDICO[64];
+extern u8 endDownEICO[64];
 
-u8 endDownGolemICO[64]={
-	1,1,25,1,1,25,1,1,
-	1,1,25,25,25,25,1,1,
-	1,1,31,31,31,31,1,1,
-	1,1,30,31,31,30,1,1,
-	1,29,30,30,30,30,29,1,
-	29,29,30,30,30,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29
-};
-u8 endDownGolemBICO[64]={
-	01,01,25,01,01,25,01,01,
-	01,01,25,25,25,25,01,01,
-	01,01,24,01,01,24,01,01,
-	01,01,24,24,24,24,01,01,
-	01,01,31,31,31,31,01,01,
-	01,8,30,31,31,30,8,01,
-	01,29,30,30,30,30,29,01,
-	01,29,30,30,30,30,29,01
-};
-u8 endDownGolemCICO[64]={
-	01,01,25,01,01,25,01,01,
-	01,01,25,25,25,25,01,01,
-	01,01,24,01,01,24,01,01,
-	01,01,24,24,24,24,01,01,
-	01,01,23,8,8,23,01,01,
-	01,8,23,23,23,23,8,01,
-	01,8,31,31,31,31,8,01,
-	01,01,30,31,31,30,01,01
-};
-u8 endDownGolemDICO[64]={
-	01,01,31,31,31,31,01,01,
-	01,01,30,31,31,30,01,01,
-	01,29,30,30,30,30,29,01,
-	29,29,30,30,30,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,01,29,30,30,29,01,29,
-	01,01,29,29,29,29,01,01,
-	01,01,29,22,22,29,01,01
-};
+extern u8 endDownGolemICO[64];
+extern u8 endDownGolemBICO[64];
+extern u8 endDownGolemCICO[64];
+extern u8 endDownGolemDICO[64];
 
-u8 golemDorflRightICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,31,8,31,8,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,29,1,1,29,1,1
-};
-u8 golemDorflRightBICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,31,8,31,8,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,1,1,1,29,1,1
-};
-u8 golemDorflRightCICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,31,8,31,8,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,29,1,1,1,1,1
-};
+extern u8 golemDorflRightICO[64];
+extern u8 golemDorflRightBICO[64];
+extern u8 golemDorflRightCICO[64];
 
-u8 golemDorflLeftICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,8,31,8,31,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,29,1,1,29,1,1
-};
-u8 golemDorflLeftBICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,8,31,8,31,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,1,1,1,29,1,1
-};
-u8 golemDorflLeftCICO[64]={
-	1,1,31,31,31,31,1,1,
-	1,1,8,31,8,31,1,1,
-	1,29,31,31,31,31,29,1,
-	29,29,30,31,31,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,1,29,30,30,29,1,29,
-	1,1,29,29,29,29,1,1,
-	1,1,29,1,1,1,1,1
-};
+extern u8 golemDorflLeftICO[64];
+extern u8 golemDorflLeftBICO[64];
+extern u8 golemDorflLeftCICO[64];
 
-u8 golemDorflUpICO[64]={
-	01,01,31,31,31,31,01,01,
-	01,01,30,31,31,30,01,01,
-	01,29,30,30,30,30,29,01,
-	29,29,30,30,30,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,01,29,30,30,29,01,29,
-	01,01,29,29,29,29,01,01,
-	01,01,29,01,01,29,01,01
-};
-u8 golemDorflUpBICO[64]={
-	01,01,31,31,31,31,01,01,
-	01,01,30,31,31,30,01,01,
-	01,29,30,30,30,30,29,01,
-	29,29,30,30,30,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,01,29,30,30,29,01,29,
-	01,01,29,29,29,29,01,01,
-	01,01,01,01,01,29,01,01
-};
-u8 golemDorflUpCICO[64]={
-	01,01,31,31,31,31,01,01,
-	01,01,30,31,31,30,01,01,
-	01,29,30,30,30,30,29,01,
-	29,29,30,30,30,30,29,29,
-	29,29,30,30,30,30,29,29,
-	29,01,29,30,30,29,01,29,
-	01,01,29,29,29,29,01,01,
-	01,01,29,01,01,01,01,01
-};
+extern u8 golemDorflUpICO[64];
+extern u8 golemDorflUpBICO[64];
+extern u8 golemDorflUpCICO[64];
 
 
-u8 zombieRightICO[64]={
-	01,01,01,01,33,33,01,01,
-	01,01,01,33,33,34,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,34,33,33,34,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,33,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,33,01,01
-};
-u8 zombieRightBICO[64]={
-	01,01,01,01,33,33,01,01,
-	01,01,01,33,33,34,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,34,33,33,34,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,33,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,33,01,01
-};
-u8 zombieRightCICO[64]={
-	01,01,01,01,33,33,01,01,
-	01,01,01,33,33,34,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,34,33,33,34,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,33,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,01,01,01
-};
+extern u8 zombieRightICO[64];
+extern u8 zombieRightBICO[64];
+extern u8 zombieRightCICO[64];
 
-u8 zombieLeftICO[64]={
-	01,01,33,33,01,01,01,01,
-	01,33,34,33,33,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,33,33,34,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,33,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,33,01,01
-};
-u8 zombieLeftBICO[64]={
-	01,01,33,33,01,01,01,01,
-	01,33,34,33,33,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,33,33,34,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,33,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,33,01,01
-};
-u8 zombieLeftCICO[64]={
-	01,01,33,33,01,01,01,01,
-	01,33,34,33,33,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,33,33,34,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,33,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,33,01,01
-};
+extern u8 zombieLeftICO[64];
+extern u8 zombieLeftBICO[64];
+extern u8 zombieLeftCICO[64];
 
-u8 zombieUpICO[64]={
-	01,01,01,33,34,01,01,01,
-	01,01,33,33,33,33,01,01,
-	01,01,33,34,33,33,01,01,
-	01,01,33,33,33,34,01,01,
-	33,01,27,27,27,27,01,33,
-	34,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,33,01,01
-};
-u8 zombieUpBICO[64]={
-	01,01,01,33,34,01,01,01,
-	01,01,33,33,33,33,01,01,
-	01,01,33,34,33,33,01,01,
-	01,01,33,33,33,34,01,01,
-	33,01,27,27,27,27,01,33,
-	34,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,01,01,01
-};
-u8 zombieUpCICO[64]={
-	01,01,01,33,34,01,01,01,
-	01,01,33,33,33,33,01,01,
-	01,01,33,34,33,33,01,01,
-	01,01,33,33,33,34,01,01,
-	33,01,27,27,27,27,01,33,
-	34,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,33,01,01
-};
+extern u8 zombieUpICO[64];
+extern u8 zombieUpBICO[64];
+extern u8 zombieUpCICO[64];
 
-u8 zombieBRightICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,33,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,35,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,33,01,01
-};
-u8 zombieBRightBICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,33,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,35,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,33,01,01
-};
-u8 zombieBRightCICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,33,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,33,33,27,27,33,35,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,01,01,01
-};
+extern u8 zombieBRightICO[64];
+extern u8 zombieBRightBICO[64];
+extern u8 zombieBRightCICO[64];
 
-u8 zombieBLeftICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,34,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,35,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,33,01,01
-};
-u8 zombieBLeftBICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,34,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,35,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,33,01,01
-};
-u8 zombieBLeftCICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,34,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,33,27,27,35,33,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,01,01,01
-};
+extern u8 zombieBLeftICO[64];
+extern u8 zombieBLeftBICO[64];
+extern u8 zombieBLeftCICO[64];
 
-u8 zombieBUpICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	35,01,27,27,27,27,01,33,
-	35,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,35,01,01
-};
-u8 zombieBUpBICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	35,01,27,27,27,27,01,33,
-	35,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,35,01,01
-};
-u8 zombieBUpCICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	35,01,27,27,27,27,01,33,
-	35,33,27,27,27,27,33,34,
-	01,01,27,27,27,27,01,01,
-	01,01,33,01,01,01,01,01
-};
+extern u8 zombieBUpICO[64];
+extern u8 zombieBUpBICO[64];
+extern u8 zombieBUpCICO[64];
 
-u8 zombieCRightICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,35,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,35,01,01
-};
-u8 zombieCRightBICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,35,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,01,01,01
-};
-u8 zombieCRightCICO[64]={
-	01,01,01,01,35,33,01,01,
-	01,01,01,35,33,35,33,01,
-	01,01,01,33,8,33,8,01,
-	01,01,01,35,33,33,35,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,35,01,01
-};
+extern u8 zombieCRightICO[64];
+extern u8 zombieCRightBICO[64];
+extern u8 zombieCRightCICO[64];
 
-u8 zombieCLeftICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,35,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,35,27,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,35,01,01
-};
-u8 zombieCLeftBICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,35,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,35,27,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,01,01,01
-};
-u8 zombieCLeftCICO[64]={
-	01,01,33,35,01,01,01,01,
-	01,33,35,33,35,01,01,01,
-	01,8,33,8,33,01,01,01,
-	01,35,33,35,33,01,01,01,
-	01,01,27,27,27,27,01,01,
-	01,35,27,27,27,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,35,01,01
-};
+extern u8 zombieCLeftICO[64];
+extern u8 zombieCLeftBICO[64];
+extern u8 zombieCLeftCICO[64];
 
-u8 zombieCUpICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,27,27,35,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,35,01,01
-};
-u8 zombieCUpBICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,27,27,35,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,01,01,01,35,01,01
-};
-u8 zombieCUpCICO[64]={
-	01,01,01,35,34,01,01,01,
-	01,01,33,33,35,33,01,01,
-	01,01,35,34,33,35,01,01,
-	01,01,33,35,33,34,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,27,27,35,27,01,01,
-	01,01,27,27,27,27,01,01,
-	01,01,35,01,01,01,01,01
-};
+extern u8 zombieCUpICO[64];
+extern u8 zombieCUpBICO[64];
+extern u8 zombieCUpCICO[64];
 
-u8 zombieDeadICO[64]={
-	01,01,01,01,01,01,01,01,
-	01,01,01,01,01,01,01,01,
-	01,01,01,01,01,01,01,01,
-	01,01,33,8,35,01,01,01,
-	01,33,35,33,33,01,01,01,
-	01,35,33,8,33,35,01,01,
-	01,01,35,33,35,35,35,01,
-	01,01,01,01,01,01,01,01
-};
+extern u8 zombieDeadICO[64];
 
 
 
@@ -1157,315 +163,45 @@ u8 zombieDeadICO[64]={
 
 
 
-u8 blueButtonUpICO[64]={
-	1,1,1,1,1,44,44,1,
-	1,1,1,1,1,44,44,1,
-	25,25,25,25,25,44,44,25,
-	25,23,23,23,23,44,44,25,
-	25,23,23,23,23,8,8,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 blueButtonMiddleICO[64]={
-	1,1,1,44,44,1,1,1,
-	1,1,1,44,44,1,1,1,
-	25,25,25,44,44,25,25,25,
-	25,23,23,44,44,23,23,25,
-	25,23,23,8,8,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 blueButtonDownICO[64]={
-	1,44,44,1,1,1,1,1,
-	1,44,44,1,1,1,1,1,
-	25,44,44,25,25,25,25,25,
-	25,44,44,23,23,23,23,25,
-	25,8,8,23,23,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 blueButtonUpICO[64];
+extern u8 blueButtonMiddleICO[64];
+extern u8 blueButtonDownICO[64];
 
-u8 doorDownBlueICO[64]={
-    1,1,1,1,1,1,1,1,
-	44,44,44,43,43,44,44,44,
-	44,44,44,43,43,44,44,44,
-	44,44,44,43,43,44,44,44,
-	44,44,44,43,43,44,44,44,
-	44,44,44,43,43,44,44,44,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownBlueBICO[64]={
-    1,1,1,1,1,1,1,1,
-	44,44,44,1,1,44,44,44,
-	44,44,44,1,1,44,44,44,
-	44,44,44,1,1,44,44,44,
-	44,44,44,1,1,44,44,44,
-	44,44,44,1,1,44,44,44,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownBlueCICO[64]={
-    1,1,1,1,1,1,1,1,
-	44,44,1,1,1,1,44,44,
-	44,44,1,1,1,1,44,44,
-    44,44,1,1,1,1,44,44,
-    44,44,1,1,1,1,44,44,
-	44,44,1,1,1,1,44,44,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorDownBlueDICO[64]={
-    1,1,1,1,1,1,1,1,
-	44,1,1,1,1,1,1,44,
-	44,1,1,1,1,1,1,44,
-    44,1,1,1,1,1,1,44,
-    44,1,1,1,1,1,1,44,
-	44,1,1,1,1,1,1,44,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 doorDownBlueICO[64];
+extern u8 doorDownBlueBICO[64];
+extern u8 doorDownBlueCICO[64];
+extern u8 doorDownBlueDICO[64];
 
 
-u8 doorLeftBlueICO[64]={
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,43,43,43,1,1,1,1,
-	1,43,43,43,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1
-};
-u8 doorLeftBlueBICO[64]={
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1
-};
-u8 doorLeftBlueCICO[64]={
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,44,44,44,1,1,1,1,
-	1,44,44,44,1,1,1,1
-};
-u8 doorLeftBlueDICO[64]={
-	1,44,44,44,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,44,44,44,1,1,1,1
-};
+extern u8 doorLeftBlueICO[64];
+extern u8 doorLeftBlueBICO[64];
+extern u8 doorLeftBlueCICO[64];
+extern u8 doorLeftBlueDICO[64];
 
-u8 doorRightBlueICO[64]={
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,43,43,43,1,
-	1,1,1,1,43,43,43,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1
-};
-u8 doorRightBlueBICO[64]={
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1
-};
-u8 doorRightBlueCICO[64]={
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,44,44,44,1
-};
-u8 doorRightBlueDICO[64]={
-	1,1,1,1,44,44,44,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,44,44,44,1
-};
+extern u8 doorRightBlueICO[64];
+extern u8 doorRightBlueBICO[64];
+extern u8 doorRightBlueCICO[64];
+extern u8 doorRightBlueDICO[64];
 
 
-u8 orangeButtonUpICO[64]={
-	1,1,1,1,1,46,46,1,
-	1,1,1,1,1,46,46,1,
-	25,25,25,25,25,46,46,25,
-	25,23,23,23,23,46,46,25,
-	25,23,23,23,23,8,8,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 orangeButtonMiddleICO[64]={
-	1,1,1,46,46,1,1,1,
-	1,1,1,46,46,1,1,1,
-	25,25,25,46,46,25,25,25,
-	25,23,23,46,46,23,23,25,
-	25,23,23,8,8,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 orangeButtonDownICO[64]={
-	1,46,46,1,1,1,1,1,
-	1,46,46,1,1,1,1,1,
-	25,46,46,25,25,25,25,25,
-	25,46,46,23,23,23,23,25,
-	25,8,8,23,23,23,23,25,
-	25,25,25,25,25,25,25,25,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 orangeButtonUpICO[64];
+extern u8 orangeButtonMiddleICO[64];
+extern u8 orangeButtonDownICO[64];
 
 
-u8 doorOrangeDownICO[64]={
-    1,1,1,1,1,1,1,1,
-	46,46,46,45,45,46,46,46,
-	46,46,46,45,45,46,46,46,
-	46,46,46,45,45,46,46,46,
-	46,46,46,45,45,46,46,46,
-	46,46,46,45,45,46,46,46,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorOrangeDownBICO[64]={
-    1,1,1,1,1,1,1,1,
-	46,46,46,1,1,46,46,46,
-	46,46,46,1,1,46,46,46,
-	46,46,46,1,1,46,46,46,
-	46,46,46,1,1,46,46,46,
-	46,46,46,1,1,46,46,46,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorOrangeDownCICO[64]={
-    1,1,1,1,1,1,1,1,
-	46,46,1,1,1,1,46,46,
-	46,46,1,1,1,1,46,46,
-	46,46,1,1,1,1,46,46,
-	46,46,1,1,1,1,46,46,
-	46,46,1,1,1,1,46,46,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
-u8 doorOrangeDownDICO[64]={
-    1,1,1,1,1,1,1,1,
-	46,1,1,1,1,1,1,46,
-	46,1,1,1,1,1,1,46,
-	46,1,1,1,1,1,1,46,
-	46,1,1,1,1,1,1,46,
-	46,1,1,1,1,1,1,46,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1
-};
+extern u8 doorOrangeDownICO[64];
+extern u8 doorOrangeDownBICO[64];
+extern u8 doorOrangeDownCICO[64];
+extern u8 doorOrangeDownDICO[64];
 
-u8 doorOrangeLeftICO[64]={
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,45,45,45,1,1,1,1,
-	1,45,45,45,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-};
-u8 doorOrangeLeftBICO[64]={
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-};
-u8 doorOrangeLeftCICO[64]={
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-};
-u8 doorOrangeLeftDICO[64]={
-	1,46,46,46,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,46,46,46,1,1,1,1,
-};
+extern u8 doorOrangeLeftICO[64];
+extern u8 doorOrangeLeftBICO[64];
+extern u8 doorOrangeLeftCICO[64];
+extern u8 doorOrangeLeftDICO[64];
 
-u8 doorOrangeRightICO[64]={
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,45,45,45,1,
-	1,1,1,1,45,45,45,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-};
-u8 doorOrangeRightBICO[64]={
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-};
-u8 doorOrangeRightCICO[64]={
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,46,46,46,1,
-};
-u8 doorOrangeRightDICO[64]={
-	1,1,1,1,46,46,46,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,1,1,1,1,
-	1,1,1,1,46,46,46,1,
-};
+extern u8 doorOrangeRightICO[64];
+extern u8 doorOrangeRightBICO[64];
+extern u8 doorOrangeRightCICO[64];
+extern u8 doorOrangeRightDICO[64];
 
-
+#endif
