+++ src/source/input.h
@@ -1,0 +1,8 @@
+#ifndef INPUT_H
+#define INPUT_H
+
+extern void waitAtoContinue();
+
+extern void waitBtoReturn();
+
+#endif
